+++ lib/mvvm_builder.dart
@@ -1,6 +1,11 @@
 import 'dart:async';
 
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
+import 'component_builder.dart';
+import 'mvvm_model.dart';
+import 'presenter_builder.dart';
 
 export 'component_builder.dart';
 export 'presenter_builder.dart';
@@ -13,4 +18,7 @@ class MvvmBuilder {
     final String version = await _channel.invokeMethod('getPlatformVersion');
     return version;
   }
+
+
 }
+