+++ lib/mvvm_context.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/cupertino.dart';
+
+class MvvmContext {
+
+  final BuildContext buildContext;
+  final AnimationController animationController;
+  final List<AnimationController> animationsControllers;
+
+  MvvmContext(this.buildContext, {this.animationController, this.animationsControllers});
+
+
+}